Fixed an issue with the getCompChoice function
When i went to create the function that would decide the winner
of the roshambo, I was getting an undefined from the getCompChoice
function. I realised i wasnt getting the answer i needed because
i hadnt told the function what to return. didnt even really need
the compChoice variable, as it's only ever being referenced
within each iteration of the winner function